pyproject.toml
[project]
name = "plane-mcp-server"
version = "0.2.8"
description = "A Model Context Protocol server for Plane integration"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name= "Surya Prashanth", email = "surya.prashanth@plane.so"},
    {name = "Plane Engineering", email = "support@plane.so"}
]
keywords = ["mcp", "plane", "fastmcp", "ai", "automation"]

dependencies = [
    "fastmcp==2.14.4",
    "plane-sdk==0.2.8",
    "py-key-value-aio[redis]==0.3.0",
    "mcp==1.26.0",
    "PyJWT>=2.12.0",
    "authlib>=1.6.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
plane-mcp-server = "plane_mcp.__main__:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["plane_mcp*"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = []

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
